+++ CHANGELOG.md
@@ -1,9 +1,11 @@
+# Unreleased
+
+### Feature
+* **terra-stopwatch** added new terra component terra-stopwatch. 
+
 <a name="3.0.0-beta.6"></a>
 # 3.0.0-beta.6 (xx.xx.2018)
-
-### Bug Fixes
 * **function-groups** added missing styles for .btn-major
-
 
 <a name="3.0.0-beta.5"></a>
 # 3.0.0-beta.5 (04.10.2018)